requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.6
django-allauth
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-htmx==1.22.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-tailwind==3.8.0
django-tinymce==4.1.0
django-tinymce4==3.0.2
django-tinymce4-lite==1.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
executing==2.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
jsmin==3.0.1
kombu==5.4.2
matplotlib-inline==0.1.7
numpy==2.2.0
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
